Cambios pre entrega

+++ src/main/java/com/example/tp/repository/AdministradorRepository.java
@@ -4,6 +4,6 @@ import com.example.tp.modelo.Administrador;
 import com.example.tp.modelo.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdministradorRepository extends JpaRepository<Administrador,Long> {
+public interface AdministradorRepository extends JpaRepository<Administrador,Integer> {
     Administrador findByDni(String dni);
 }
